Finish refactoring
Stop appending to the same list if more than one 'I am creating a list'
case, because it's super-duper confusing.

Even though it can be useful if you want a mix of things in that list.

Hope no breakage :(
