+++ main.go
@@ -1,23 +1,16 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"os"
-	"os/exec"
-	"path"
-	"path/filepath"
 	"strings"
 
 	"github.com/imdario/mergo"
 	jsone "github.com/taskcluster/json-e"
-	jsone_interpreter "github.com/taskcluster/json-e/interpreter"
 	// Quick hack of ghodss YAML to expose a new method
 	yaml_ghodss "github.com/wryun/yaml-1"
 	yaml_v2 "gopkg.in/yaml.v2"
@@ -69,201 +62,6 @@ type context interface {
 	eval() (interface{}, error)
 }
 
-type commonContext struct {
-	key string
-	raw bool
-}
-
-func (cc *commonContext) cleanupResult(result interface{}) (interface{}, error) {
-	var finalResult interface{}
-	if cc.raw {
-		switch r := result.(type) {
-		case []byte:
-			finalResult = string(r)
-		default:
-			// otherwise, let's just let JSON-e complain...
-			finalResult = result
-		}
-	} else {
-		var rawBytes []byte
-		switch r := result.(type) {
-		case string:
-			rawBytes = []byte(r)
-		case []byte:
-			rawBytes = r
-		default:
-			return nil, errors.New("somehow got a non string/byte slice?")
-		}
-		if err := yaml_ghodss.Unmarshal(rawBytes, &finalResult); err != nil {
-			return nil, err
-		}
-	}
-
-	if cc.key != "" {
-		return map[string]interface{}{cc.key: finalResult}, nil
-	}
-
-	return finalResult, nil
-}
-
-type fileContext struct {
-	commonContext
-	filename string
-}
-
-func (fc *fileContext) eval() (interface{}, error) {
-	result, err := ioutil.ReadFile(fc.filename)
-	if err != nil {
-		return nil, err
-	}
-
-	return fc.cleanupResult(result)
-}
-
-type stdinContext struct {
-	commonContext
-	filename string
-}
-
-func (sc *stdinContext) eval() (interface{}, error) {
-	result, err := ioutil.ReadAll(os.Stdin)
-	if err != nil {
-		return nil, err
-	}
-
-	return sc.cleanupResult(result)
-}
-
-type functionContext struct {
-	commonContext
-	function  string
-	rawOutput bool
-	rawInput  bool
-}
-
-func (fc *functionContext) eval() (interface{}, error) {
-	var f interface{}
-	commandArray := strings.Split(fc.function, " ")
-
-	if fc.rawInput && fc.rawOutput {
-		f = func(args []interface{}, stdin string) (string, error) {
-			stringArgs, err := castToStrings(args)
-			extendedCommandArray := append(commandArray, stringArgs...)
-			command := exec.Command(extendedCommandArray[0], extendedCommandArray[1:]...)
-			command.Stderr = os.Stderr
-			command.Stdin = bytes.NewReader([]byte(stdin))
-			stdoutBytes, err := command.Output()
-			if err != nil {
-				return "", err
-			}
-			return string(stdoutBytes), nil
-		}
-	} else if fc.rawInput {
-		f = func(args []interface{}, stdin string) (interface{}, error) {
-			stringArgs, err := castToStrings(args)
-			extendedCommandArray := append(commandArray, stringArgs...)
-			command := exec.Command(extendedCommandArray[0], extendedCommandArray[1:]...)
-			command.Stderr = os.Stderr
-			command.Stdin = bytes.NewReader([]byte(stdin))
-			stdoutBytes, err := command.Output()
-			if err != nil {
-				return nil, err
-			}
-
-			var o interface{}
-			err = yaml_ghodss.Unmarshal(stdoutBytes, &o)
-			if err != nil {
-				return nil, err
-			}
-			return o, nil
-		}
-	} else if fc.rawOutput {
-		f = func(args []interface{}, stdin interface{}) (string, error) {
-			jsonBytes, err := json.Marshal(stdin)
-			if err != nil {
-				return "", err
-			}
-
-			stringArgs, err := castToStrings(args)
-			extendedCommandArray := append(commandArray, stringArgs...)
-			command := exec.Command(extendedCommandArray[0], extendedCommandArray[1:]...)
-			command.Stderr = os.Stderr
-			command.Stdin = bytes.NewReader(jsonBytes)
-			stdoutBytes, err := command.Output()
-			return string(stdoutBytes), err
-		}
-	} else {
-		f = func(args []interface{}, stdin interface{}) (interface{}, error) {
-			jsonBytes, err := json.Marshal(stdin)
-			if err != nil {
-				return "", err
-			}
-
-			stringArgs, err := castToStrings(args)
-			extendedCommandArray := append(commandArray, stringArgs...)
-			command := exec.Command(extendedCommandArray[0], extendedCommandArray[1:]...)
-			command.Stderr = os.Stderr
-			command.Stdin = bytes.NewReader(jsonBytes)
-			stdoutBytes, err := command.Output()
-			if err != nil {
-				return nil, err
-			}
-
-			var o interface{}
-			err = yaml_ghodss.Unmarshal(stdoutBytes, &o)
-			if err != nil {
-				return nil, err
-			}
-			return stdin, nil
-		}
-	}
-
-	return jsone_interpreter.WrapFunction(f), nil
-}
-
-type textContext struct {
-	commonContext
-	text string
-}
-
-func (tc *textContext) eval() (interface{}, error) {
-	return tc.cleanupResult(tc.text)
-}
-
-type listContext struct {
-	commonContext
-	contexts []context
-	metadata bool
-}
-
-func (lc *listContext) eval() (interface{}, error) {
-	outputList := make([]interface{}, 0, len(lc.contexts))
-
-	for _, context := range lc.contexts {
-		result, err := context.eval()
-		if err != nil {
-			return nil, err
-		}
-
-		if !lc.metadata {
-			outputList = append(outputList, result)
-		}
-
-		metadataResult := map[string]interface{}{
-			"content": result,
-		}
-		if fc, ok := context.(*fileContext); ok {
-			basename := path.Base(fc.filename)
-			metadataResult["filename"] = fc.filename
-			metadataResult["basename"] = basename
-			metadataResult["name"] = strings.TrimSuffix(basename, filepath.Ext(basename))
-		}
-		outputList = append(outputList, metadataResult)
-	}
-
-	return lc.cleanupResult(outputList)
-}
-
 func main() {
 	var args arguments
 	flag.Usage = func() {
@@ -285,66 +83,6 @@ func main() {
 	if err := run(logger, os.Stdout, args); err != nil {
 		fmt.Fprintf(flag.CommandLine.Output(), "Fatal error: %s\n", err)
 		os.Exit(2)
-	}
-}
-
-func parseContexts(rawContextInfo []string) []context {
-	contexts := make([]context, 0)
-
-	currentContexts := &contexts
-
-	for _, contextOp := range rawContextInfo {
-		key := ""
-		var entry string
-		if strings.HasPrefix(contextOp, "+") {
-			entry = contextOp
-		} else {
-			splitContextOp := strings.SplitN(contextOp, ":", 2)
-			if len(splitContextOp) < 2 {
-				entry = splitContextOp[0]
-			} else {
-				key = splitContextOp[0]
-				entry = splitContextOp[1]
-				// If we have a new key, we should jump out of any list we're in
-				currentContexts = &contexts
-			}
-		}
-
-		context := parseContext(key, entry)
-		*currentContexts = append(*currentContexts, context)
-		if listContext, ok := context.(*listContext); ok {
-			currentContexts = &listContext.contexts
-		}
-	}
-
-	return contexts
-}
-
-func parseContext(key string, entry string) context {
-	baseContext := commonContext{
-		raw: strings.HasPrefix(entry, ":"),
-		key: key,
-	}
-
-	if baseContext.raw {
-		entry = entry[1:]
-	}
-
-	switch {
-	case entry == "..":
-		return &listContext{commonContext: baseContext, metadata: false}
-	case entry == "...":
-		return &listContext{commonContext: baseContext, metadata: true}
-	case strings.HasPrefix(entry, "+"):
-		return &textContext{commonContext: baseContext, text: entry[1:]}
-	case entry == "-":
-		return &stdinContext{commonContext: baseContext}
-	case strings.HasPrefix(entry, "--"):
-		return &functionContext{commonContext: commonContext{raw: true, key: key}, rawInput: baseContext.raw, rawOutput: true, function: entry[2:]}
-	case strings.HasPrefix(entry, "-"):
-		return &functionContext{commonContext: commonContext{raw: true, key: key}, rawInput: baseContext.raw, rawOutput: false, function: entry[1:]}
-	default:
-		return &fileContext{commonContext: baseContext, filename: entry}
 	}
 }
 
@@ -430,7 +168,7 @@ func run(l *log.Logger, out io.Writer, args arguments) error {
 func loadContext(contexts []context, deepMerge bool) (map[string]interface{}, error) {
 	finalContext := make(map[string]interface{})
 
-	for _, context := range contexts {
+	for i, context := range contexts {
 		untypedNewContext, err := context.eval()
 		if err != nil {
 			return nil, err
@@ -438,7 +176,7 @@ func loadContext(contexts []context, deepMerge bool) (map[string]interface{}, er
 
 		newContext, ok := untypedNewContext.(map[string]interface{})
 		if !ok {
-			return nil, errors.New("TODO cannot have unkeyed top level context")
+			return nil, fmt.Errorf("context at position %d had no top level keys: %q", i, untypedNewContext)
 		}
 
 		if deepMerge {
@@ -454,16 +192,4 @@ func loadContext(contexts []context, deepMerge bool) (map[string]interface{}, er
 	}
 
 	return finalContext, nil
-}
-
-func castToStrings(slice []interface{}) ([]string, error) {
-	result := make([]string, len(slice))
-	for i, v := range slice {
-		if s, ok := v.(string); !ok {
-			return nil, errors.New("function command line arguments must be strings (use stdin or $json)")
-		} else {
-			result[i] = s
-		}
-	}
-	return result, nil
 }
